requirements.txt
# qrtools==0.0.2
pyzbar==0.1.9
pillow==10.4.0